+++ sputofy_2/lib/screens/playlistListScreen/playlist_list_screen.dart
@@ -1,11 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:sputofy_2/models/playlist_model.dart';
-import 'package:sputofy_2/models/playlist_song_model.dart';
-import 'package:sputofy_2/models/song_model.dart';
 import 'package:sputofy_2/providers/provider.dart';
-import 'package:sputofy_2/screens/playlistSongsScreen/playlist_songs_screen.dart';
-import 'package:sputofy_2/theme/palette.dart';
+
+import 'components/playlist_tile.dart';
 
 class PlaylistListScreen extends StatefulWidget {
   const PlaylistListScreen({Key? key}) : super(key: key);
@@ -23,7 +21,7 @@ class _PlaylistListScreenState extends State<PlaylistListScreen> {
           List<Playlist> playlists = database.playlists;
 
           return Column(
-            children: [
+            children: <Widget>[
               PlaylistList(context, playlists),
             ],
           );
@@ -48,111 +46,35 @@ class PlaylistList extends StatelessWidget {
         itemCount: playlists.length,
         itemBuilder: (context, index) {
           Playlist playlist = playlists[index];
-          return Tooltip(
-            message: playlist.name,
-            padding: const EdgeInsets.all(16.0),
-            child: Card(
-              elevation: 4.0,
-              color: kSecondaryBackgroundColor,
-              shape: RoundedRectangleBorder(
-                borderRadius: BorderRadius.all(
-                  Radius.circular(10.0),
-                ),
-              ),
-              child: InkWell(
-                onTap: () => Navigator.push(
-                  context,
-                  MaterialPageRoute(
-                    builder: (context) => PlaylistSongsScreen(playlist),
-                  ),
-                ),
-                child: Row(
-                  mainAxisSize: MainAxisSize.max,
-                  children: <Widget>[
-                    Stack(
-                      alignment: AlignmentDirectional.center,
-                      children: <Widget>[
-                        ClipRRect(
-                          borderRadius: BorderRadius.only(
-                              topLeft: Radius.circular(10.0),
-                              bottomLeft: Radius.circular(10.0)),
-                          child: Image.asset(
-                            'cover.jpeg',
-                            width: 70.0,
-                            height: 70.0,
-                          ),
-                        ),
-                        IconButton(
-                            splashColor: Colors.transparent,
-                            iconSize: 48.0,
-                            onPressed: () => print("bottone"),
-                            icon: Icon(
-                              Icons.play_arrow,
-                              color: Colors.white,
-                              // size: 48.0,
-                            )),
-                      ],
-                    ),
-                    SizedBox(width: 16.0),
-                    Expanded(
-                      child: Column(
-                        crossAxisAlignment: CrossAxisAlignment.start,
-                        children: <Widget>[
-                          Text(
-                            playlist.name,
-                            style: Theme.of(context).textTheme.subtitle1,
-                            overflow: TextOverflow.ellipsis,
-                          ),
-                          FutureBuilder<List<Song>>(
-                              future: Provider.of<DBProvider>(context,
-                                      listen: false)
-                                  .testGetPlaylistSongs(playlist.id!),
-                              builder: (context, snapshot) {
-                                if (!snapshot.hasData)
-                                  return CircularProgressIndicator();
-                                return Text(
-                                  snapshot.data?.length.toString() ?? '0',
-                                  style: Theme.of(context).textTheme.subtitle2,
-                                );
-                              })
-                        ],
-                      ),
-                    ),
-                    // SizedBox(width: 8.0),
-                    _buildWidgetMenuButton(playlist),
-                  ],
-                ),
-              ),
-            ),
-          );
+          return PlaylistTile(playlist);
         },
       ),
     );
   }
+}
 
-  Widget _buildWidgetMenuButton(Playlist playlist) {
-    return PopupMenuButton<List>(
-      onSelected: _handleClick,
-      icon: Icon(Icons.more_vert),
-      padding: EdgeInsets.zero,
-      itemBuilder: (context) {
-        return {'Delete Playlist'}.map((String choice) {
-          return PopupMenuItem<List>(
-            value: [choice, playlist],
-            child: Text(choice),
-          );
-        }).toList();
-      },
-    );
+void handleClick(List params, BuildContext context) {
+  //* params = [choice, playlist]
+  switch (params[0]) {
+    case 'Delete Playlist':
+      Provider.of<DBProvider>(context, listen: false)
+          .deletePlaylist(params[1].id);
+      break;
   }
+}
 
-  void _handleClick(List params) {
-    //* params = [choice, playlist]
-    switch (params[0]) {
-      case 'Delete Playlist':
-        Provider.of<DBProvider>(context, listen: false)
-            .deletePlaylist(params[1].id);
-        break;
-    }
-  }
+Widget PlaylistbuildWidgetMenuButton(Playlist playlist, BuildContext context) {
+  return PopupMenuButton<List>(
+    onSelected: (List params) => handleClick(params, context),
+    icon: Icon(Icons.more_vert),
+    padding: EdgeInsets.zero,
+    itemBuilder: (context) {
+      return {'Delete Playlist'}.map((String choice) {
+        return PopupMenuItem<List>(
+          value: [choice, playlist],
+          child: Text(choice),
+        );
+      }).toList();
+    },
+  );
 }
